🎨 automated cleanup

+++ src/Core/ArgumentOutOfRangeExceptionThrower.ThrowIfNegative.cs
@@ -36,7 +36,7 @@ public partial class ArgumentOutOfRangeExceptionEx
     /// <exception cref="ArgumentOutOfRangeException"><paramref name="value"/> is negative.</exception>
     public static void ThrowIfNegative(sbyte value, [Runtime.CompilerServices.CallerArgumentExpression(nameof(value))] string? paramName = null)
     {
-        const short Zero = (sbyte)0;
+        const sbyte Zero = 0;
         if (value < Zero)
         {
             ThrowNegative(value, paramName);
@@ -51,7 +51,7 @@ public partial class ArgumentOutOfRangeExceptionEx
     /// <exception cref="ArgumentOutOfRangeException"><paramref name="value"/> is negative.</exception>
     public static void ThrowIfNegative(short value, [Runtime.CompilerServices.CallerArgumentExpression(nameof(value))] string? paramName = null)
     {
-        const short Zero = (short)0;
+        const short Zero = 0;
         if (value < Zero)
         {
             ThrowNegative(value, paramName);
